Added the material ui, and started to create the first page
I'm trying to create a whole page first, then I will decouple in several react components and organize. The reason I'm doing this is because I'm not familiarized with React Components and their props and states. Therefore, I will first worry with the design and get things working, then I will decouple in components. The last step is to connect with the API